Integrate 8.1 pre with main (#18)
* Update form Python2 to Python3. (Note old mbuild still uses python2)

* Bugfixing python3 scripts, crucial for building

* Update to PYTHON3 scripts with some bugfixing

* Bug-fix for Python3 support

* Fixes for Python 3.12

* Update to spawning the files with the Python interpreter

* Hotfix, of deadlock with GIL when releasing the simulator

* Bugfix causes a deadlock in Python 3.12 at finalize

* Logic fix

* Removed some useless csv

* Update Dockerfile requirements

* Added Dockerfile for Ubuntu 24.04

* Rebasing issue in another branch

* Update to undo regression with commit 632a6493108243f6b334c0744253702ef77f6a4e

* Update to use SDE as the default, with a version from 9.7 to 9.38.
Some bugfixes were necessary for the CPUID emulation, where (probably)
SDE overwrote some of the registers after manually changing it.
There is a bug in libc of LLVM that prevents inputstream to work
properly, this is fixed by using the c-api fread instead.
Lastly an update to python script, to gracefully shut down the
front-end.

* Support for c++20, updated XED and PIN (not PINPLAY)

* Sniper can now compile with latest gcc/g++

* Fixing some c++20 compilation issues

* Regressing some commits, for compilation to succeed

* Little fix, to stop if the thread it was not stopped already

* Use the warmup-start event of Pin/SDE to switch to cache-only mode

* Added a flag  to emulate a specific microarchitecture (Ex: snb, bdw, skx, adl) on the front-end

* Front-end support for the latest SDE v9.44 kit

* Add libtorch and neural-net based branch predictor (#16)

* Add libtorch to MakeFile and MakeFile.common

* Update gitignore

* Add neural net branch predictor

* Temporarily add libtorch to ld path

* Use c++20 instead of c++2a

* Add torch_home to config.py and ld path

* Rename semaphore.h to sem.h to avoid conflicts in std

* Resolve circular dependancy

* Fixed libtorch version to 2.5.0

* Add fft as test

* Change download link to snipersim.org

* Add fft test for nn branch predictor


---------

+++ scripts/dvfs.py
@@ -15,7 +15,7 @@ class Dvfs:
     self.events = []
     args = args.split(':')
     for i in range(0, len(args), 3):
-      self.events.append((long(args[i])*sim.util.Time.NS, int(args[i+1]), int(args[i+2])))
+      self.events.append((int(args[i])*sim.util.Time.NS, int(args[i+1]), int(args[i+2])))
     self.events.sort()
     sim.util.Every(100*sim.util.Time.NS, self.periodic, roi_only = True)
